7228: Moved logger to core
