+++ assets/client/core/logger.js
