+++ assets/client/service/token-service.js
@@ -1,4 +1,4 @@
-import logger from "../logger";
+import logger from "../core/logger.js";
 import appStorage from "../core/app-storage";
 import ClientConfigLoader from "../core/client-config-loader.js";
 import defaults from "../util/defaults";
